requirements.txt
numpy
pandas
torch
sentence-transformers
faiss-cpu
hazm
scikit-learn
joblib